backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
opencv-python==4.11.0.86
websockets==15.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
python-dotenv==0.9.0
minio==7.2.15
confluent-kafka[avro, schemaregistry]==2.10.0
lap==0.5.12
python-jose==3.5.0
pydantic[email]==2.11.3
passlib==1.7.4
aiohttp==3.12.12
weasyprint==65.1